+++ src/generate-report.ts
@@ -3,12 +3,34 @@ import * as path from 'path';
 import { compile, registerHelper } from 'handlebars';
 import { CoverageDetails } from './model';
 
+/**
+ *
+ * @param detail A `CoverageDetails` object returned from the `parseLcov` method. These are the values that will be filled into the reports.
+ * @param outputPath The path to where you would like the files written to.
+ * @param mediumThreshold The percentage (0-1) of when cells should move from red to yellow. Defaults to 0.75.
+ * @param highThreshold The percentage (0-1) of when cells should move from yellow to green. Defaults to 1.0.
+ */
 export function generateReport(
   detail: CoverageDetails,
   outputPath: string,
   mediumThreshold = 0.75,
   highThreshold = 1
 ): void {
+  if (mediumThreshold < 0 || mediumThreshold > 1) {
+    throw new Error(
+      `mediumThreshold should be between 0 and 1, but was ${mediumThreshold}`
+    );
+  }
+  if (highThreshold < 0 || highThreshold > 1) {
+    throw new Error(
+      `highThreshold should be between 0 and 1, but was ${highThreshold}`
+    );
+  }
+  if (mediumThreshold >= highThreshold) {
+    throw new Error(
+      `mediumThreshold should be strictly less than highThreshold.`
+    );
+  }
   const roundToTwoDecimals = (num: number) =>
     Math.round((num + Number.EPSILON) * 10000) / 100;
 
@@ -43,11 +65,11 @@ export function generateReport(
   });
 
   const mainTemplateString = fse.readFileSync(
-    path.join(__dirname, '../../templates/main_report_template.handlebars'),
+    path.join(__dirname, '../templates/main_report_template.handlebars'),
     'utf-8'
   );
   const fileTemplateString = fse.readFileSync(
-    path.join(__dirname, '../../templates/file_report_template.handlebars'),
+    path.join(__dirname, '../templates/file_report_template.handlebars'),
     'utf-8'
   );