+++ src/model/branch_coverage_details.ts
@@ -1,6 +1,0 @@
-export interface BranchCoverageDetails {
-  line: number;
-  block: number;
-  branch: number;
-  taken: number;
-}