+++ src/model/file_coverage_details.ts
@@ -1,4 +1,4 @@
-import { LineCoverageDetails } from '.';
+import { FunctionCoverageDetails, LineCoverageDetails } from '.';
 
 /**
  * Parsed details from an indidual file, including
@@ -9,6 +9,8 @@ export interface FileCoverageDetails {
   path: string;
   /** Coverage details for each particular line. */
   lines: LineCoverageDetails[];
+  /** Coverage details for each particular function. */
+  functions: FunctionCoverageDetails[];
   /** Total lines hit in this file. */
   linesHit: number;
   /** Total lines found in this file. */
@@ -21,4 +23,6 @@ export interface FileCoverageDetails {
   functionsHit: number;
   /** Total functions found in this file. */
   functionsFound: number;
+  /** If true, this file should be excluded from code coverage. */
+  excluded: boolean;
 }