+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+# 0.0.1-dev.2
+
+- feat: you can now exclude files when using `lcovParse`
+- docs: every object and function now has documentation
+- docs: created auto-generated the [documentation site](https://mtwichel.github.io/very-good-coverage-core) for detailed documentation
+
 # 0.0.1-dev.1
 
-- feat: lcovParse function
-- feat: generateReport function
+- feat: `lcovParse` function
+- feat: `generateReport` function